AudioManager + Enemy Stun
Created an AudioManager and implemented it in the MainMenu Scene.

Modified the enemy script to have a public Stun() function which is called when it gets hit so that it can apply a knockback.
This was needed because linearVelocity is being overridden every frame in MoveTowards() when chasing the player.